PD-14397 add screener story
