+++ packages/SidePanel/stories/SidePanel.Tests.stories.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Screener from "./examples/Screener";
+
+export default {
+  title: " | SidePanel/ Automation Tests",
+};
+
+export const screener = () => <Screener />;
